Project/requirements.txt
asgiref==3.5.2
Django==4.0.4
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
psycopg2===2.9.3
psycopg2-binary==2.9.3